seems to be working well with this release.  Fixed some minor bugs.

+++ include/LogDistributer.h
@@ -27,7 +27,7 @@ private:
     LogDistributerAnalyzer *mp_analyzer;
     //int m_numberOfArchivers;
 
-    int64_t m_index;
+    int m_index;
 
     bool m_closed;
